backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
xmltodict
langchain
langchain-google-genai
chromadb
python-dotenv
PyPDF2
langchain-community
langchain-chroma
langchain-text-splitters
langchain-core
python-multipart
PyJWT
bcrypt
SQLAlchemy
